requirements.txt
Flask==2.3.3
Flask-MySQL==1.5.2
mysql-connector-python==8.1.0
matplotlib==3.7.2
werkzeug==2.3.7